Tidied code more and moved module imports, need to organise these for all code base
